cleanup
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@9345 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/packages/xpath/cc/fetch_file.c
@@ -38,7 +38,7 @@ int get_file_www(char *server, char * fname, char **buf);
 int parse_url( const char * url, char * server, char *fname)
 {
   int i,j;
-  int len = strlen(url);
+  int len = (int)strlen(url);
   char temp[MAXSTRLEN];
   
   int flag = 0, flag_file = 0;
